refactor: 패키지 구조 정리 - 예약
