+++ src/test/java/roomescape/ReservationsTest.java
@@ -7,6 +7,8 @@ import java.time.LocalDate;
 import java.time.LocalTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.usecase.Reservation.Reservation;
+import roomescape.usecase.Reservation.Reservations;
 
 public class ReservationsTest {
